requirements.txt
Flask==3.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.1
openpyxl==3.1.2